refactor: database for artist
