+++ src/artist/artist.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { ArtistService } from './artist.service';
 import { ArtistController } from './artist.controller';
-import { InMemoryDB } from 'src/InMemoryDB';
+import { InMemoryDB } from 'src/helpers/InMemoryDB';
 
 @Module({
   controllers: [ArtistController],
