+++ src/artist/artist.service.ts
@@ -3,11 +3,11 @@ import { v4 as uuidv4 } from 'uuid';
 import { CreateArtistDto } from './dto/create-artist.dto';
 import { UpdateArtistDto } from './dto/update-artist.dto';
 import { Artist } from './entities/artist.entity';
-import { InMemoryDB } from 'src/InMemoryDB';
+import { InMemoryDB } from 'src/helpers/InMemoryDB';
 
 @Injectable()
 export class ArtistService {
-  constructor(private readonly inMemoryDB: InMemoryDB) {}
+  constructor(private readonly inMemoryDB: InMemoryDB<Artist>) {}
 
   create(createArtistDto: CreateArtistDto): Artist {
     const artist = {
@@ -15,50 +15,37 @@ export class ArtistService {
       id: uuidv4(),
     };
 
-    this.inMemoryDB.artists.push(artist);
+    this.inMemoryDB.create(artist);
     return artist;
   }
 
   findAll(): Artist[] {
-    return this.inMemoryDB.artists;
+    return this.inMemoryDB.findAll();
   }
 
   findOne(id: string): Artist {
-    const artist = this.inMemoryDB.artists.find((artist) => artist.id === id);
+    const artist = this.inMemoryDB.findOne(id);
 
-    if (artist) {
-      return artist;
-    } else {
+    if (!artist) {
       throw new NotFoundException(`Artist with id ${id} not found`);
     }
+
+    return artist;
   }
 
   update(id: string, updateArtistDto: UpdateArtistDto): Artist {
-    const index = this.inMemoryDB.artists.findIndex(
-      (artist) => artist.id === id,
-    );
+    const artist = this.findOne(id);
 
-    if (index === -1) {
-      throw new NotFoundException(`Artist with id ${id} not found`);
-    } else {
-      return (this.inMemoryDB.artists[index] = {
-        ...this.inMemoryDB.artists[index],
-        ...updateArtistDto,
-      });
-    }
+    return this.inMemoryDB.update(artist, { ...updateArtistDto, id });
   }
 
   remove(id: string): string {
-    const length = this.inMemoryDB.artists.length;
+    const isRemoved = this.inMemoryDB.delete(id);
 
-    this.inMemoryDB.artists = this.inMemoryDB.artists.filter(
-      (artist) => artist.id !== id,
-    );
-
-    if (length === this.inMemoryDB.artists.length) {
+    if (isRemoved) {
       throw new NotFoundException(`Artist with id ${id} not found`);
-    } else {
-      return `Artist with id: ${id} removed`;
     }
+
+    return `Artist with id: ${id} removed`;
   }
 }
